Add `ChromicPDF.OutputOptions`
In order to have a single module where we manage how to output PDF files, we have created `ChromicPDF.OutputOptions` (using plural as we also do it for `PDFOptions`).

In here we can find the previous `feed_chrome_data_into_output/2` and a new `feed_file_into_output/2` that replaces `PDFAOptions.feed_ghostscript_file_into_output/2` . That's because we will also use the same function to output merged PDF files.
